+++ src/ClassicUO.Client/LegionScripting/docs/API.py
@@ -2263,13 +2263,14 @@ def HasGump(ID: "int" = 1337) -> "int":
     """
     pass
 
-def ReplyGump(button: "int", gump: "int" = 1337, switches: "list[int]" = None) -> "bool":
+def ReplyGump(button: "int", gump: "int" = 1337, switches: "list[int]" = None, entries: "list[Any]" = None) -> "bool":
     """
      Reply to a gump.
      Example:
      ```py
      API.ReplyGump(21)
      API.ReplyGump(1, 0x555, [100])
+     API.ReplyGump(1, 0x555, [100], [(0, "text input")])
      ```
     
     """